Attempting a different algorithm for rbyd range removals
The previous attempt to make range removals more rigorous highlighted a
pretty significant design flaw: Every removals risks making the tree ~2x
taller.

In theory this is offset by the fact that removals, well, remove nodes,
shrinking the height of the tree, but this isn't reflected in the
underlying red-black-yellow structure. Blanket recoloring breaks the
red-black-yellow invariants.

This isn't the end of the world, we still rebalance during rbyd
compaction, but it would be nice if we had stronger guarantees about
the structure of rbyds before compaction. Especially since we rely on
tree balance to defend our O(n log n) traversal overhead.

---

This attempts to reimplement range removals with two separate passes for
diverged trunk.

The downside is range removals now need, well, two separate passes, even
if we don't actually end up with a diverged trunk. The upside is in
theory we can preserve the coloring information and related invariants.

I would describe this commit as "almost working" and "a mess". There
still seems to be some issues with null tags getting stuck in the tree
after diverged trunks.

On the plus side, our tests are certainly working...