Fixed diverged leaf coloring issue
Took some debugging to figure out what was going on, but this was just
a refactoring oversight, didn't update the diverged state used to decide
how to color the leaves.

I guess this would have actually been caught earlier if I cleaned up the
code before debugging the test failures. But I wanted to reach
proof-of-concept first...

Anyways, good news, all tests are passing, so the proof-of-concept
new range removal algorithm works.