Cleaned up reworked range removal/diverging trunk algorithm
Note this is still blanket recoloring alts to black in diverged trunks.
So we don't get the main benefit of this rework yet: preserving the rbyd
color invariants.

But this is a nice checkpoint that shows that our two-pass diverging
trunk algorithm works without breaking right-leaning invariants.

Before, we made a single pass, and stitched together alternating
alts in any diverged trunks in order to remove ranges of tags (ignore
coloring for now, coloring _does_ help here):

                                          .-----------o
         .-------o-------.                |           o-----------.
     .---o---.       .---o---.            |     .-----o           |
   .-o-.   .-o-.   .-o-.   .-o-.        .-o-.   |     o-----.   .-o-.
  .o. .o. .o. .o. .o. .o. .o. .o.      .o. .o. .o. .--o--. .o. .o. .o.
  a b c d e f g h i j k l m n o p  =>  a b c d e f g     j k l m n o p
               '-+-'
               remove

Now, we do two passes: One to write any alts on the lower-diverged
trunk. And one to write the non-diverging part of the trunk, stitch in the
lower-diverged trunk with an alt, and then write any alts on the
upper-diverged trunk.

This results in a somewhat complex diverging state machine:

  diverge possible?  diverge not possible?
         |                    |
         v                    |
   DIVERGINGLOWER-----------. |
         |                  | |
         v                  v v
   DIVERGEDLOWER        NOTDIVERGING
         |                   |
         v                   |
   DIVERGINGUPPER            |
         |                   |
         v                   |
   DIVERGEDUPPER             |
         '--------. .--------'
                  v v
               leaf stuff

But the end result is a much more balanced tree, at least on first
inspection:

         .-------o-------.                         .--o--.
     .---o---.       .---o---.            .--------o     o--------.
   .-o-.   .-o-.   .-o-.   .-o-.        .-o-.   .--o     o--.   .-o-.
  .o. .o. .o. .o. .o. .o. .o. .o.      .o. .o. .o. |     | .o. .o. .o.
  a b c d e f g h i j k l m n o p  =>  a b c d e f g     j k l m n o p
               '-+-'
               remove

Unfortunately, this does mean we need, well, two passes, even if the
resulting tree doesn't actually end up with a diverging trunk. We can
avoid two passes if we know a diverged trunk is impossible, but we only
know this in the single attr append case (no negative deltas, no
submask, no supmask, etc).

At the very least we don't end up _progging_ any more alts than is
necessary. Even though we write two trunks, the number of alts after
pruning end up the same as in our single-pass algorithm. The only
additional cost is a single null tag (4 bytes) to terminate the diverged
trunk.

This two pass algorithm may sound more complicated, and therefore more
costly, but keep in mind this replaces the previous diverged-swapping
state, which I would argue resulted in more mess and a harder to
understand lfsr_rbyd_appendattr.

Now that the dust has settled, we can actually start comparing the
relative code costs, though again note we are still blanket recoloring.
The result is a surprising basically net-zero code code:

             code          stack
  one-pass: 33852           2880
  two-pass: 33860 (+0.0%)   2880 (+0.0%)