Found a balance-preserving solution to tail-recursive range recoloring
It feels a bit clumsy, but by using an additional bit of state to keep
track of if the last alt was pruned, we can cancel recolorings that may
risk recursion.

If we look at how this plays out on the underlying 2-3-4 tree:

   .-----.        .-------.         .-------.           .-------.
   |.a.h.|        |.a.c.h.|         |.a.c.h.|           |.a.c.h.|
   '|-|-|'        '|-|-|-|'         '|-|-|-|'           '|-|-|-|'
      |            .-' '-.          .--' '--.            .-' '-.
      v            v     v          v       v            v     v
  .-------.      .---. .---.      .---. .-------.      .---. .---.
  |.b.c.g.|  =>  |.b.| |.g.|  =>  |.b.| |.d.e.f.|  =>  |.b.| |.e.|
  '|-|-|-|'      '|-|' '|-|'      '|-|' '|-|-|-|'      '|-|' '|-|'
       | x              | x                                 .-' '-.
       v                v                                   v     v
   .-------.        .-------.                             .---. .---.
   |.d.e.f.|        |.d.e.f.|                             |.d.| |.f.|
   '|-|-|-|'        '|-|-|-|'                             '|-|' '|-|'

Note the important property that no nodes ended up at a height _worse_
than where they started.

It's interesting to note this is equivalent to splitting the nodes
_before_ prunning:

   .-----.        .-------.        .-------.          .-------.
   |.a.h.|        |.a.c.h.|        |.a.c.h.|          |.a.c.h.|
   '|-|-|'        '|-|-|-|'        '|-|-|-|'          '|-|-|-|'
      |            .-' '-.          .-' '--.           .-' '-.
      v            v     v          v      v           v     v
  .-------.      .---. .---.      .---. .-----.      .---. .---.
  |.b.c.g.|  =>  |.b.| |.g.|  =>  |.b.| |.e.g.|  =>  |.b.| |.e.|
  '|-|-|-|'      '|-|' '|-|'      '|-|' '|-|-|'      '|-|' '|-|'
       | x              | x          .---' | x            .-' '-.
       v                v            v     v              v     v
   .-------.        .-------.      .---. .---.          .---. .---.
   |.d.e.f.|        |.d.e.f.|      |.d.| |.f.|          |.d.| |.f.|
   '|-|-|-|'        '|-|-|-|'      '|-|' '|-|'          '|-|' '|-|'

Which is probably why most of our 2-3-4 tree invariants hold.

In the actual implementation, we encode the current pruned state as a
part of our diverging state machine, since we don't non-trivially prune
outside of diverging trunks.

This ends up with the following, slightly-extended, diverging state
machine:

  diverge possible?          diverge not possible?
         |                            |
         v                            |
   DIVERGINGLOWER-------------------. |
         |                          | |
         v                          v v
   DIVERGEDLOWER<->PRUNEDLOWER  NOTDIVERGING
         | .------------'            |
         v v                         |
   DIVERGINGUPPER                    |
         |                           |
         v                           |
   DIVERGEDUPPER<->PRUNEDUPPER       |
         '------------. | .----------'
                      v v v
                    leaf stuff

Writing out the state machine like this actually highlights the slightly
annoying transition from PRUNEDUPPER to leaf stuff, which was buggy in
the first impl.

We also encode some common information (lower/upper, pruned, etc) in the
state machine's bit encoding to try to avoid too many if statements.
Though this impl does seem a bit heavy handed.

The additional complexity results in of course more code cost, but as a
trade-off our range recoloring should be a bit more sturdy and provably
preserves the h=2log2(b) worst case height of our tree:

                          code          stack
  broken recoloring:     33880           2880
  unbalanced recoloring: 33912 (+0.1%)   2880 (+0.0%)
  balanced recoloring:   33944 (+0.2%)   2880 (+0.0%)