Tweaked name/sizelimit parsing for consistency, default to 0xff/0x7fffffff
This is mostly just for consistency with changes to parsing other parts
of the fs config attrs.

This changes name/size limits to default to namelimit=0xff and
sizelimit=0x7fffffff. These are reasonable defaults for 32-bit systems,
which was the original use case for littlefs. Though, with the diversity
of embedded device, I suspect these will be overridden more often than
not. For this reason the 0xff/0x7fffffff case is not treated specially
during lfs_format and these limits are always written. Though this may
change in the future.

The intention behind these defaults is to align with other limits that
may be introduced in the future. Any new artificial limits will
necessarily require defaulting to their existing values for backwards
compatibility, so hopefully this allows all limits to be handled
consistently.

If a future use-case-specific implementation of littlefs can benefit
from assuming these defaults, that's a nice plus.

Name/size-limit attr encodings:

  .---+---+---+---.      tag (0x000c): 1 be16    2 bytes
  | x000c | 0 |siz|      weight (0):   1 leb128  1 byte
  +---+---+---+---+      size:         1 leb128  1 byte
  | name_limit    |      name_limit:   1 leb128  <=4 bytes
  '---+- -+- -+- -'      total:                  <=8 bytes

  .---+---+---+---.      tag (0x000d): 1 be16    2 bytes
  | x000d | 0 |siz|      weight (0):   1 leb128  1 byte
  +---+---+---+---+- -.  size:         1 leb128  1 byte
  | size_limit        |  size_limit:   1 leb128  <=5 bytes
  '---+- -+- -+- -+- -'  total:                  <=9 bytes

Code changes:

           code          stack
  before: 34040           2880
  after:  34060 (+0.1%)   2880 (+0.0%)