Added type info to dsize comments
This is mostly just a lot of leb128s, though we do use be16 for tags and
le32 for cksums and revision counts.

There are several places we use single-byte leb128s, which really are
u8s with the top bit reserved. Still, notating this as leb128 indicates
that the top bit really is reserved, even if you don't need full leb128
encoding/decoding in practice.