Don't duplicate config across all mroots, only magic
So, duplicating the config across multiple mroots allows a better
chance of manual recovery if things go wrong, right?

   .--------.  .--------.
  .|littlefs|->|littlefs|
  ||bs=4096 | ||bs=4096 |
  ||bc=256  | ||bc=256  |
  ||crc32c  | ||root dir|
  ||        | ||crc32c  |
  |'--------' |'--------'
  '--------'  '--------'

Well, this was the original thinking. But now I'm starting to think the
duplicated config isn't actually all that useful:

1. The config may be out-of-date, since only the last mroot is mutable.

   This is becoming more common as littlefs evolves (on-disk version
   bumps, compat flag changes, fs_grow, etc).

2. The most important bit of information is where the mtree is. And this
   information is _only_ available on the last mroot because of
   mutability requirements.

   At the very least, all mroots point to the mtree (not Rome). So
   finding the mtree may not be that hard if you find any mroot. But
   this also gives you all the config sooo...

3. gstate is going to be hard (impossible?) to reconstruct anyways.
   Though for reading this may only be an issue for interrupted grms.

4. Let's be honest, manual recovery is not going to be a common
   occurence for these devices.

Point 1. is a the main issue and actually highlights a real risk with
duplicated config: It's easy to pick up out-of-date config.

In other implementations this risks easy mistakes that are hard to
notice until complex filesystem states. Assuming the first mroot
contains up-to-date config for the obvious example.

Even in our current implementation, duplicated config already poses some
tricky hard-to-get-right problems. What happens if we run into an
unknown compat flag on a not-last mroot? Hint, our implementation was
broken!

---

So this commit changes mroot extension to _not_ duplicate config, but
instead just rewrite the magic string and mroot chain to the new mroot
anchor:

   .--------.  .--------.
  .|littlefs|->|littlefs|
  ||crc32c  | ||bs=4096 |
  ||        | ||bc=256  |
  ||        | ||root dir|
  ||        | ||crc32c  |
  |'--------' |'--------'
  '--------'  '--------'

This leads to a nice simplification in lfsr_mdir_commit, so that's a
plus.

It does make lfsr_mount a bit trickier, since we need to figure out which
mroot is the last mroot before checking for config. But we would need
something trickier in lfsr_mount anyways to handle the above out-of-date
issues anyways.

The current implementation just does a redundant mroot lookup to figure
out if the current mroot is the last mroot. In theory this could be
avoided, but I couldn't figure out how to without making the code
unreasonably complex (lfsr_mount is already intertwined with
lfsr_traversal_read).

The end result is a bit of code and stack savings, thanks to
lfsr_mdir_commit being on the stack-depth hot-path (deep-path?):

           code          stack
  before: 34060           2880
  after:  33996 (-0.2%)   2864 (-0.6%)

It's also worth noting that there are plans to add block-level
redundancy at some point. Maybe it's best to leave recovering from
missing blocks to block-level redundancy which is actually designed for
this, and let the mroot chain do what the mroot chain does best:
allowing the mroots to participate in wear-leveling.