Replace assert(!err) with assert(err == 0) in tests
This plays better with prettyasserts.py, which prints the err value on
failure.

We _could_ extend prettyasserts.py to print the contents of !err
patterns, but this risks making the error message more confusing when
the target is an actual boolean expression. Keep in mind
prettyasserts.py is purely syntactical and doesn't really know the
expression's type.