rbyd-rr: Made remove leaf-splits red
This avoids the alta prune cludge, where we prune altas unconditionally
knowing we only emit these to make removes work.

The alta prune cludge was a bit concerning forward-compatibility-wise,
since it technically violates the rby structure of the tree, but
necessary to prevent unbalancing when we terminate remove leaves with
black altas. Terminating with a black alta technically also violates the
rby structure, and two wrongs make a right, right?

But why are these altas black? To be honest it's just what made the code
work in the moment. These should be red, but terminating with red altas
turned out to be surprisingly tricky.

The problem is when we terminate with a red alta, the alta is subject to
recoloring, and may be reordered as a part of a yellow node to preserve
the yellow-alts-point-same-dir invariant. But if you reorder the alta,
anything after it becomes unreachable! Not good!

  altrgt 0x200           altrgt 0x200
  altrle 0x100    =>     altra
  altba         yellow   altble 0x100 <-- unreachable!
  null          reorder  null

The solution here turned out to just not use altas at all. If we're
careful with our tag bounds, we can create an alt that is _implicitly_
alta without a special encoding. Such an alt can be reordered without
issue:

  altrgt 0x200           altrgt 0x200
  altrle 0x100    =>     altrgt 0x100
  altbgt 0x100  yellow   altble 0x100 <-- reachable
  null          reorder  null         <-- unreachable

The other option would have been to make lfsr_rbyd_p_recolor alta aware,
but this would have been quite complicated and fully of special cases...

The "if we're careful with our tag bounds" is the tricky bit, since we
didn't really need to be that careful before. But the end result is
tracking diverged tag bounds the same way we track diverged rid bounds,
which is a nice bit of consistency. This also avoids annoying yellow
terminating d_state corner cases. It's a nice improvement.

As a part of these changes I also tweaked to the lower_tag bound to
track last seen alt instead of alt+1. To be honest I'm not really sure
how alt+1 got there. I guess to be consistent with the upper_tag bound?
Bound the upper_tag bound is exclusive, so this ends up weird and
difficult to reason about...

This change also makes it so _all_ leaf splits end up red, which is
unexpected but nice for consistency. We can probably make leaf-split
recoloring unconditional eventually.