rbyd-rr: Fixed unreachable red alts not being pruned
It turns out we were never pruning unreachable red alts. I thought we
were because of lfsr_tag_prune2 and related logic, but this was custom
tailored for the specific unreachable patterns created by yellow splits,
and is insufficient for all unreachable alts.

Consider this unreachable altbgt 0x300 (the second one):

  altbgt 0x300 ----------> altbgt 0x300
  altrle 0x200 -----.----> altrgt 0x300 <-- unreachable,
  altbgt 0x300 -----'----> altble 0x200     should have
  null             =>      tag 0x100        been pruned
                 append
                tag 0x100

Our prune logic doesn't catch this because the altbgt is pointing a
different direction than the altrle we end up taking.

This wasn't an issue for earlier range-removal algorithms, since we have
a separate explicit check for diverged pruning to avoid weight ambiguity
issues. Black altas were also not an issue because this logic does catch
unreachable black alts, which are a bit easier. But now that we are
emitting intentionally unreachable red alts with the expectation that
they will be cleaned up by our pruning logic, this is a bit of a
problem...

The solution here is to check for unreachable alts after red flips. This
duplicates quite a bit of code but avoids the logical complexity of
figuring out reachability in all the permutations of red 2-3 nodes.